Cleanup for the new VIDEO cnf keyword, corrected some types, added range checking to the cnf parser.

+++ xaaes/src.km/menuwidg.c
@@ -48,7 +48,7 @@ new_task(Tab *new)
 }
 
 static void
-free_task(Tab *t, int *nester)
+free_task(Tab *t, short *nester)
 {
 	if (t)
 	{
